refactor(cli): extract manifest OAuth domain
